+++ fwdpp/sugar/metapop.hpp
@@ -31,7 +31,7 @@ namespace KTfwd
 				 boost::container::vector<diploid_t>,
 				 boost::container::vector<boost::container::vector<diploid_t>>,
 				 boost::container::vector<mtype>,
-				 boost::container::vector<unsigned>,
+				 boost::container::vector<uint_t>,
 				 boost::unordered_set<double,boost::hash<double>,KTfwd::equal_eps>>;
 
   /*!
@@ -52,7 +52,7 @@ namespace KTfwd
 						       boost::container::vector<diploid_t>,
 						       boost::container::vector<boost::container::vector<diploid_t>>,
 						       boost::container::vector<mtype>,
-						       boost::container::vector<unsigned>,
+						       boost::container::vector<uint_t>,
 						       boost::unordered_set<double,boost::hash<double>,KTfwd::equal_eps>,
 						       diploid_writer_t,
 						       diploid_reader_t>;
@@ -83,7 +83,7 @@ namespace KTfwd
 				 std::vector<diploid_t>,
 				 std::vector<std::vector<diploid_t> >,
 				 std::vector<mtype>,
-				 std::vector<unsigned>,
+				 std::vector<uint_t>,
 				 std::unordered_set<double,std::hash<double>,KTfwd::equal_eps>>;
 
   /*!
@@ -104,7 +104,7 @@ namespace KTfwd
 						       std::vector<diploid_t>,
 						       std::vector<std::vector<diploid_t> >,
 						       std::vector<mtype>,
-						       std::vector<unsigned>,
+						       std::vector<uint_t>,
 						       std::unordered_set<double,std::hash<double>,KTfwd::equal_eps>,
 						       diploid_writer_t,
 						       diploid_reader_t>;