+++ fwdpp/sugar/sampling/sampling_details.hpp
@@ -7,8 +7,6 @@ namespace KTfwd
   {
     enum class treat_neutral {ALL,NEUTRAL,SELECTED};
   }
-  using sample_t = std::vector< std::pair<double,std::string> >;
-  using sep_sample_t = std::pair<sample_t,sample_t>;
 
   template<typename vec_mutation_t>
   void add_fixations( sample_t * sample,
@@ -92,8 +90,8 @@ namespace KTfwd
       }
     auto rv = std::move(temp.first);
     std::move(temp.second.begin(),temp.second.end(),std::back_inserter(rv));
-    std::sort(rv.begin(),rv.end(),[](const std::pair<double,std::string> & a,
-				     const std::pair<double,std::string> & b){
+    std::sort(rv.begin(),rv.end(),[](const sample_site_t & a,
+				     const sample_site_t & b){
 		return a.first<b.first;
 	      });
     if(!removeFixed)
@@ -115,8 +113,8 @@ namespace KTfwd
       }
     auto rv = std::move(temp.first);
     std::move(temp.second.begin(),temp.second.end(),std::back_inserter(rv));
-    std::sort(rv.begin(),rv.end(),[](const std::pair<double,std::string> & a,
-				     const std::pair<double,std::string> & b){
+    std::sort(rv.begin(),rv.end(),[](const sample_site_t & a,
+				     const sample_site_t & b){
 		return a.first<b.first;
 	      });
     if(!removeFixed)