+++ fwdpp/sugar/singlepop.hpp
@@ -54,7 +54,7 @@ namespace KTfwd
 							   singlepop_glist_t<mtype>,
 							   boost::container::vector< diploid_t >,
 							   boost::container::vector<mtype>,
-							   boost::container::vector<unsigned>,
+							   boost::container::vector<uint_t>,
 							   boost::unordered_set<double,boost::hash<double>,KTfwd::equal_eps>,
 							   diploid_writer_t,
 							   diploid_reader_t
@@ -87,7 +87,7 @@ namespace KTfwd
 				       singlepop_glist_t<mtype>,
 				       std::vector< diploid_t >,
 				       std::vector<mtype>,
-				       std::vector<unsigned>,
+				       std::vector<uint_t>,
 				       std::unordered_set<double,std::hash<double>,KTfwd::equal_eps>
 				       >;
 
@@ -109,7 +109,7 @@ namespace KTfwd
 							   singlepop_glist_t<mtype>,
 							   std::vector< diploid_t >,
 							   std::vector<mtype>,
-							   std::vector<unsigned>,
+							   std::vector<uint_t>,
 							   std::unordered_set<double,std::hash<double>,KTfwd::equal_eps>,
 							   diploid_writer_t,
 							   diploid_reader_t